requirements.txt
openai>=1.30.0
langgraph>=0.2.0
langchain-core>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
aiofiles>=24.1.0
httpx>=0.27.0
python-multipart>=0.0.12
websockets>=13.0
rich>=13.9.0
structlog>=24.4.0
tenacity>=9.0.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
pypdf>=4.0.0
python-docx>=1.1.0
